Fixes #25882.

+++ code/game/machinery/kitchen/gibber.dm
@@ -21,6 +21,12 @@
 	. = ..()
 	update_icon()
 
+/obj/machinery/gibber/Destroy()
+	if(occupant)
+		occupant.dropInto(loc)
+		occupant = null
+	. = ..()
+
 /obj/machinery/gibber/on_update_icon()
 	overlays.Cut()
 	if (dirty)
@@ -190,22 +196,22 @@
 
 	admin_attack_log(user, occupant, "Gibbed the victim", "Was gibbed", "gibbed")
 	src.occupant.ghostize()
+	addtimer(CALLBACK(src, .proc/finish_gibbing), gib_time)
 
-	spawn(gib_time)
+/obj/machinery/gibber/proc/finish_gibbing()
+	operating = 0
+	if(QDELETED(occupant))
+		occupant = null
+		return
+	occupant.gib()
+	qdel(occupant)
 
-		src.operating = 0
-		src.occupant.gib()
-		qdel(src.occupant)
-
-		playsound(src.loc, 'sound/effects/splat.ogg', 50, 1)
-		operating = 0
-		for (var/obj/thing in contents)
-			// There's a chance that the gibber will fail to destroy some evidence.
-			if(istype(thing,/obj/item/organ) && prob(80))
-				qdel(thing)
-				continue
-			thing.dropInto(loc) // Attempts to drop it onto the turf for throwing.
-			thing.throw_at(get_edge_target_turf(src,gib_throw_dir),rand(0,3),emagged ? 100 : 50) // Being pelted with bits of meat and bone would hurt.
-		update_icon()
-
-
+	playsound(loc, 'sound/effects/splat.ogg', 50, 1)
+	for (var/obj/thing in (contents - component_parts))
+		// There's a chance that the gibber will fail to destroy some evidence.
+		if(istype(thing,/obj/item/organ) && prob(80))
+			qdel(thing)
+			continue
+		thing.dropInto(loc) // Attempts to drop it onto the turf for throwing.
+		thing.throw_at(get_edge_target_turf(src,gib_throw_dir),rand(0,3),emagged ? 100 : 50) // Being pelted with bits of meat and bone would hurt.
+	update_icon()
